백 데이터 api 테스트 및 오타 수정

+++ src/client/client.service.ts
@@ -16,11 +16,14 @@ import { ClientsInput } from './dtos/clients.input';
 import { OrderEnum } from 'src/common/dtos/find-many.input';
 import * as ExcelJS from 'exceljs';
 import { FindDateInput } from 'src/common/dtos/find-date.input';
-import { FilterQuery } from 'mongoose';
+import { FilterQuery, Model } from 'mongoose';
+import { InjectModel } from '@nestjs/mongoose';
+import { Sale } from 'src/sale/entities/sale.entity';
 
 @Injectable()
 export class ClientService {
   constructor(
+    @InjectModel(Sale.name) private readonly saleModel: Model<Sale>,
     private readonly clientRepository: ClientRepository,
     private readonly utilService: UtilService,
     private readonly saleService: SaleService,
@@ -49,6 +52,42 @@ export class ClientService {
           `${input.name}는 이미 사용중인 이름 입니다.`,
         );
       }
+    }
+  }
+
+  private async beforeUpdate(input: UpdateClientInput) {
+    if (input.code) {
+      throw new BadRequestException('거래처 코드는 수정할 수 없습니다.');
+    }
+
+    if (input.name) {
+      const isSaleRecordExist = await this.saleModel.exists({
+        mallId: input.name,
+      });
+
+      if (isSaleRecordExist) {
+        throw new BadRequestException(
+          `${input.name} 판매기록이 있는 제품이름 입니다. 사방넷 전산과 동기화를 위해 이름을 수정할 수 없습니다.`,
+        );
+      }
+
+      const isNameExist = await this.clientRepository.exists({
+        _id: { $ne: input._id },
+        name: input.name,
+      });
+
+      if (isNameExist) {
+        throw new BadRequestException(
+          `${input.name}는 이미 사용중인 이름 입니다.`,
+        );
+      }
+    }
+
+    const updateTarget = await this.clientRepository.model.find({
+      _id: input._id,
+    });
+    if (!updateTarget) {
+      throw new BadRequestException('해당 거래처를 찾을 수 없습니다.');
     }
   }
 
@@ -87,11 +126,26 @@ export class ClientService {
   }
 
   async update({ _id, ...body }: UpdateClientInput) {
-    await this.beforeCreate({ ...body, _id });
+    await this.beforeUpdate({ ...body, _id });
     return this.clientRepository.update({ _id }, body);
   }
 
   async remove(_id: string) {
+    const deleteTarget = await this.clientRepository.findOne({ _id });
+    if (!deleteTarget) {
+      throw new BadRequestException('해당 거래처를 찾을 수 없습니다.');
+    }
+
+    const hasSaleRecord = await this.saleModel.exists({
+      mallId: deleteTarget.name,
+    });
+
+    if (hasSaleRecord) {
+      throw new BadRequestException(
+        '판매 기록이 있는 거래처는 삭제할 수 없습니다.',
+      );
+    }
+
     const result = await this.clientRepository.remove({ _id });
     return result;
   }
